+++ e2e-tests/components/proposals-tab.ts
@@ -1,4 +1,5 @@
 import { MyNavigator } from "../common/navigator";
+import { execFile } from "node:child_process";
 
 export class ProposalsTab extends MyNavigator {
   static readonly SELECTOR: string = `[data-tid="proposals-tab"]`;
@@ -7,6 +8,48 @@ export class ProposalsTab extends MyNavigator {
 
   public static proposalIdSelector(proposalId: number): string {
     return `[data-tid="proposal-id"][data-proposal-id="${proposalId}"]`;
+  }
+
+  /**
+   * Creates a proposal via the command line tool.
+   */
+  public static async propose(proposalName: string): Promise<number> {
+    console.error("Creating proposal...");
+    const proposalId: number = await new Promise((yay, nay) =>
+      execFile(
+        "../scripts/propose",
+        ["--to", proposalName, "--jfdi"],
+        {},
+        (error, stdout, stderr) => {
+          console.error("stdout:", stdout);
+          console.error("stderr:", stderr);
+          if (error) {
+            nay(new Error(`${error}\nstdout: ${stdout}\nstderr: ${stderr}`));
+          } else {
+            const proposalId = Number(
+              stdout
+                .split(/[\n\r]/)
+                .map((line) => line.split(/[ \t]+/))
+                .filter(
+                  (fields) => fields.length === 2 && fields[0] === "proposal"
+                )
+                .map((fields) => fields[1])[0]
+            );
+            if (Number.isSafeInteger(proposalId)) {
+              yay(proposalId);
+            } else {
+              nay(
+                new Error(
+                  `No proposal was made.\nstdout:\n${stdout}\nstderr: ${stderr}`
+                )
+              );
+            }
+          }
+        }
+      )
+    );
+    console.error(`Proposed proposalId ${proposalId}`);
+    return proposalId;
   }
 
   constructor(browser: WebdriverIO.Browser) {
